updating file

+++ 7. Number Theory/3. Binary Exponentiation/3. Large Exponentiation.cpp
@@ -1,8 +1,8 @@
 #include<bits/stdc++.h>
 using namespace std;
-const int M = 1e18 + 7;
-int binMultiply(long long a, long  long b){
-    int ans = 0;
+const long long M = 1e18 + 7;
+long long binMultiply(long long a, long  long b){
+    long long ans = 0;
     while(b){
         if(b&1){
             ans = (ans + a) % M;
@@ -12,8 +12,8 @@ int binMultiply(long long a, long  long b){
     }
     return ans;
 }
-int binExpIter(int a, int b){
-    int ans = 1;
+long long binExpIter(long long a, long long b){
+    long long ans = 1;
     while(b){
         if(b&1){
             ans = binMultiply(ans,a);
@@ -25,7 +25,7 @@ int binExpIter(int a, int b){
 }
 //TC Log^2(n) cause 2 log(n) loops
 int main(){
-    int a, b;
+    long long a, b;
     cin >> a >> b;
     cout << binExpIter(a,b);
 }
